requested changes implemented; createElement,changeView funcs moved to Controller folder
