+++ js/main.js
@@ -1,4 +1,1 @@
 import './templates/index';
-import './Controller/pointCounter';
-import './Controller/timer';
-import './data/generateTestResults';
