+++ js/templates/game3.js
@@ -1,5 +1,5 @@
-import createElement from '../createElement';
-import changeView from '../changeView';
+import createElement from '../Controller/createElement';
+import changeView from '../Controller/changeView';
 import {stats, intro} from './index';
 
 const template = `<div><header class="header">
